Moved tests to subdirectory, updated test makefile and fixed apikey dictionary reference.
